Refactored optuna objectives
